Audio issue fixed, Release Candidate 1
The audio issue has been patched successfully in this commit.

Other changes:
OGG Vorbis audio conversion now not only works, but is 5 times faster-by
switching to libvorbis

A brand new interface which keeps processes safe and makes bug-fixing
easier

With this build, ConvertX is back to Beta, and not just Beta-this is now
a Release Candidate.

Current version: 0.3 Beta Release Candidate 1